requirements.txt
azure-identity==1.17.1
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.21.0
azure-storage-queue==12.8.0
applicationinsights==0.11.10
beautifulsoup4==4.12.3
langdetect==1.0.9
lxml==5.2.2
numpy==1.26.4
pandas==2.2.2
pydantic==1.10.15
requests==2.32.3
sentence-transformers==2.7.0
scikit-learn==1.5.1
hdbscan==0.8.33
transformers==4.44.2
accelerate==0.33.0
pytest==8.2.2
flake8==7.1.1
black==24.8.0

